reformulating the package and todim is done
