+++ examples/example_a_topsis.py
@@ -1,0 +1,6 @@
+import sys
+sys.path.append("../src")
+from decision_making import TODIM
+
+td = TODIM([[1,2], [1,2]])
+print(td.matrix_d)
