+++ src/decision_making/__init__.py
@@ -1,0 +1,1 @@
+from decision_making.todim import TODIM
