+++ src/decision_making/a_topsis.py
@@ -25,7 +25,6 @@ from __future__ import division
 import sys
 sys.path.insert(0, '../topsis')
 
-from methods import topsis
 import numpy as np
 
 # The matrix of means
